+++ packages/bindings/src/profiles/querier.rs
@@ -6,7 +6,7 @@ use crate::iter::page_iterator::{Page, PageIterator};
 use cosmwasm_std::Binary;
 
 use crate::profiles::types::*;
-use crate::types::PageRequest;
+use crate::cosmos_types::PageRequest;
 use cosmwasm_std::{Addr, Empty, QuerierWrapper, StdResult};
 
 /// Querier allows to query data from the Desmos x/profiles module.
