+++ packages/bindings/src/reactions/querier.rs
@@ -9,7 +9,7 @@ use crate::{
 use cosmwasm_std::Binary;
 
 use crate::reactions::types::*;
-use crate::types::PageRequest;
+use crate::cosmos_types::PageRequest;
 use cosmwasm_std::{Addr, Empty, QuerierWrapper, StdResult};
 
 /// Querier allows to query data from the Desmos x/reactions module.
